Componentize bridge parts
